+++ layout/includes/footnote.php
@@ -15,18 +15,24 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Theme Boost Campus - Version file
+ * Theme Boost Campus - Layout file for footnote.
  *
- * @package    theme_boost_campus
- * @copyright  2017 Kathrin Osswald, Ulm University <kathrin.osswald@uni-ulm.de>
- * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ * @package   theme_boost_campus
+ * @copyright 2017 Kathrin Osswald, Ulm University kathrin.osswald@uni-ulm.de
+ * @copyright based on code from theme_boost by Damyon Wiese
+ * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
 defined('MOODLE_INTERNAL') || die();
 
-$plugin->component = 'theme_boost_campus';
-$plugin->version = 2017060400;
-$plugin->release = 'v3.2-r1';
-$plugin->requires = 2016070700;
-$plugin->maturity = MATURITY_STABLE;
-$plugin->dependencies = array('theme_boost' => 2016102100);
+$footnotesetting = get_config('theme_boost_campus', 'footnote');
+
+// Only proceed if text area does not only contains empty tags.
+if (!html_is_blank($footnotesetting)) {
+    // Use format_string function to enable multilanguage filtering.
+    $footnotesetting = format_text($footnotesetting);
+
+    $templatecontext['footnotesetting'] = $footnotesetting;
+
+    echo $OUTPUT->render_from_template('theme_boost_campus/footnote', $templatecontext);
+}