+++ lib.php
@@ -88,6 +88,7 @@ function theme_boost_campus_get_pre_scss($theme) {
         'branddangercolor' => ['brand-danger'],
         'darknavbar' => ['darknavbar'],
         'footerblocks' => ['footerblocks'],
+        'badgeareaitemsmaxheight' => ['badgeareaitemsmaxheight'],
         /* MODIFICATION END */
     ];
 
@@ -133,12 +134,27 @@ function theme_boost_campus_pluginfile($course, $cm, $context, $filearea, $args,
             return $theme->setting_file_serve('favicon', $args, $forcedownload, $options);
         } else if ($filearea === 'loginbackgroundimage') {
             return $theme->setting_file_serve('loginbackgroundimage', $args, $forcedownload, $options);
-         } else if ($filearea === 'fontfiles') {
+        } else if ($filearea === 'fontfiles') {
             return $theme->setting_file_serve('fontfiles', $args, $forcedownload, $options);
-         } else {
+        } else if ($filearea === 'badgeareaitems') {
+            return $theme->setting_file_serve('badgeareaitems', $args, $forcedownload, $options);
+        } else {
             send_file_not_found();
         }
     } else {
         send_file_not_found();
     }
+}
+
+/**
+ * If setting is updated, use this callback to clear the theme_boost_campus' own application cache.
+ */
+function theme_boost_campus_reset_app_cache() {
+    // Get the cache from area.
+    $themeboostcampuscache = cache::make('theme_boost_campus', 'badgearea');
+    // Delete the cache for the badgearea.
+    $themeboostcampuscache->delete('badgeareadata');
+    // To be safe and because there can only be one callback function added to a plugin setting,
+    // we also delete the complete theme cache here.
+    theme_reset_all_caches();
 }