+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2017-06-04 - Settings to be able to add additional regions beneath the page footer.
 * 2017-05-22 - Added some section separator lines to the footer.
 * 2017-05-22 - Adding the possibility to give additional classes to the list group items of the flat navigation.
 * 2017-05-16 - Changed naming and reference of mustache template that renders the login form.